refactor: extract config error handling
